TH-1051: fix ts problems
