+++ babel.config.json
@@ -3,6 +3,5 @@
   "plugins": [
     "@babel/proposal-class-properties",
     "@babel/proposal-object-rest-spread",
-
   ]
 }
