+++ load-tests/k6-utils.ts
@@ -4,7 +4,7 @@ import { parseHTML, Selection } from 'k6/html';
 import http from 'k6/http';
 
 //TODO import does not work?
-//import { ROUTES } from '../src/domain/app/routes/constants';
+import { ROUTES } from '../src/domain/app/routes/constants';
 
 // Environment variables can be added to run command:
 // eslint-disable-next-line max-len
@@ -14,25 +14,6 @@ import http from 'k6/http';
 export const BASE_URL =
   // eslint-disable-next-line no-undef
   __ENV.BASE_URL || 'https://tapahtumat.test.kuva.hel.ninja';
-
-export const GRAPHQL_BASE_URL =
-  // eslint-disable-next-line no-undef
-  __ENV.GRAPHQL_BASE_URL ||
-  'https://tapahtumat-proxy.test.kuva.hel.ninja/proxy/graphql';
-
-const ROUTES = {
-  ABOUT: '/about',
-  ACCESSIBILITY: '/accessibility',
-  COLLECTION: '/collection/:slug',
-  COLLECTIONS: '/collections',
-  EVENTS: '/events',
-  EVENT: '/events/:id',
-  EVENT_DEPRECATED: '/event/:id',
-  COURSES: '/courses',
-  COURSE: '/courses/:id',
-  HOME: '/home',
-  HOME_PREVIEW: '/home/:id',
-};
 
 export type PageType = {
   [K in keyof typeof ROUTES]: typeof ROUTES[K] | ((param: string) => string);
