+++ package.json
@@ -76,8 +76,8 @@
     "serialize-javascript": "^3.1.0",
     "typescript": "3.9.3",
     "use-deep-compare-effect": "^1.4.0",
-    "webpack": "4.43.0",
-    "webpack-cli": "^3.3.11",
+    "webpack": "4.42.0",
+    "webpack-cli": "^3.3.12",
     "xml-js": "^1.6.11"
   },
   "scripts": {
@@ -118,14 +118,9 @@
     ]
   },
   "devDependencies": {
-    "@babel/cli": "^7.8.4",
-    "@babel/core": "^7.8.4",
-    "@babel/node": "^7.8.4",
+    "@apollo/react-testing": "^3.1.3",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/plugin-proposal-object-rest-spread": "^7.8.3",
-    "@babel/preset-env": "^7.8.4",
-    "@babel/preset-typescript": "^7.8.3",
-    "@apollo/react-testing": "^3.1.3",
     "@ffmpeg-installer/ffmpeg": "^1.0.20",
     "@graphql-codegen/cli": "^1.9.1",
     "@graphql-codegen/typescript": "1.9.1",
