+++ src/domain/course/CoursePageContainer.tsx
@@ -19,7 +19,8 @@ import EventClosedHero from '../event/eventClosedHero/EventClosedHero';
 import EventContent from '../event/eventContent/EventContent';
 import EventHero from '../event/eventHero/EventHero';
 import EventPageMeta from '../event/eventPageMeta/EventPageMeta';
-import { getEventIdFromUrl, isEventClosed } from '../event/EventUtils';
+import { isEventClosed } from '../event/EventUtils';
+import { getEventIdFromUrl } from '../event/EventUrlUtils';
 import { useSimilarCoursesQuery } from '../event/queryUtils';
 import SimilarEvents from '../event/similarEvents/SimilarEvents';
 import { EventFields, SuperEventResponse } from '../event/types';
