+++ src/domain/event/EventUtils.ts
@@ -18,7 +18,7 @@ import {
   EVENT_PLACEHOLDER_IMAGES,
   EVENT_SOME_IMAGE,
 } from './constants';
-import { EventFields, EventType, KeywordOption } from './types';
+import { EventFields, KeywordOption } from './types';
 
 export const getEventCardId = (id: string): string => `event-card_${id}`;
 
@@ -52,30 +52,6 @@ export const isEventFree = (event: EventFieldsFragment): boolean => {
   const offer = event.offers.find((item) => item.isFree);
 
   return !!offer?.isFree;
-};
-
-/**
- * Get event id from url
- * For example  https://api.hel.fi/linkedcourses/v1/event/harrastushaku:13433?query -> harrastushaku:13433
- */
-export const getEventIdFromUrl = (
-  url: string,
-  type: EventType = 'event'
-): string | undefined => {
-  return url.match(new RegExp(`/(?:${type}s?)/([^/?]*)`, 'i'))?.[1];
-};
-
-export const getEventIdsFromUrls = (
-  urls: string[]
-): { eventIds: string[]; courseIds: string[] } => {
-  return {
-    eventIds: urls
-      .map((url) => getEventIdFromUrl(url, 'event') as string)
-      .filter(Boolean),
-    courseIds: urls
-      .map((url) => getEventIdFromUrl(url, 'course') as string)
-      .filter(Boolean),
-  };
 };
 
 /*
