+++ src/domain/event/__tests__/EventUtils.test.ts
@@ -11,11 +11,10 @@ import {
   fakeKeywords,
   fakePlace,
 } from '../../../test/mockDataUtils';
+import { getEventIdFromUrl, getEventIdsFromUrls } from '../EventUrlUtils';
 import {
   formatPrice,
   getEventFields,
-  getEventIdFromUrl,
-  getEventIdsFromUrls,
   getEventImageUrl,
   getEventPlaceholderImageUrl,
   getEventSomeImageUrl,
